+++ CHANGELOG.md
@@ -4,13 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.4.3] - 2021-04-28
 ### Fixed
 - Removed spurious start_gap and end_gap in TraceTransition.
 - Missing layer in PositiveCPWBlank.
 
 ### Added
 - TraceTransitionBlank, a placeholder that draws nothing.
+- PositiveCPWTransitionDummy, a placeholder that draws nothing.
+- individual_keywords in `SegmentList.draw` to allow different keywords to be passed to different Segments.
 
 ## [0.4.2] - 2021-04-16
 ### Changed